Ignore facterdb default datasets (#51)
Before this patch, a call to an os missing from the simp factsets would
fall through to facterdb's default data.  However, this data is
inappropriate for SIMP spec tests―it is missing critical facts and the
keys are no deep-symbolized

For the time being, this patch simply ignores non-simp factsets by not
merging results from facterdb.  This is quick measure to get simp/simp
spec tests working with missing factsets from 3.6.0 (OEL7, RHEL7)

A follow-on patch is expected to remove all logic to to with lookups
from default facterdb data sources